requirements.txt
asgiref==3.8.1
dj-static==0.0.6
Django==5.1.1
django-stdimage==6.0.2
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
static3==0.7.0
